methoden ausgelagert
